+++ _test_projects/basics/bundle/index.py
@@ -1,5 +1,8 @@
 # cob: type=views mountpoint=/bundle/index
 from cob import route
+from cob.project import get_project
+
+from flask import jsonify
 
 @route('/')
 def index():