+++ cob/subsystems/manager.py
@@ -4,7 +4,7 @@ import yaml
 import logbook
 
 from .base import SubsystemBase
-from ..exceptions import MountpointConflict
+from ..exceptions import MountpointConflict, UnknownSubsystem
 from ..utils.parsing import parse_front_matter
 
 _logger = logbook.Logger(__name__)
@@ -33,7 +33,7 @@ class SubsystemsManager(object):
                 continue
             subsystems = {grain.subsystem for grain in grains}
             if len(subsystems) > 1 or not grains[0].subsystem.SUPPORTS_OVERLAYS:
-                raise MountpointConflict('Mount point {} used more than once'.format(mountpoint))
+                raise MountpointConflict(f'Mount point {mountpoint} used more than once')
 
     def _load_project_subsystems(self):
         roots = [self.project.root]
@@ -55,7 +55,10 @@ class SubsystemsManager(object):
                     _logger.trace('Will traverse into bundle {}', path)
                     roots.append(path)
                     continue
-                subsystem_cls = self._get_subsystem_by_grain_type(grain_type)
+                try:
+                    subsystem_cls = self._get_subsystem_by_grain_type(grain_type)
+                except UnknownSubsystem:
+                    raise UnknownSubsystem(f'Grain {path} uses an unknown subsystem type: {grain_type!r}') from None
                 subsystem = self._subsystems.get(subsystem_cls.NAME)
                 if subsystem is None:
                     subsystem = self._subsystems[
@@ -84,7 +87,10 @@ class SubsystemsManager(object):
             subsystem.configure_app(flask_app)
 
     def _get_subsystem_by_grain_type(self, grain_type):
-        return SubsystemBase.SUBSYSTEM_BY_NAME[grain_type]
+        try:
+            return SubsystemBase.SUBSYSTEM_BY_NAME[grain_type]
+        except KeyError:
+            raise UnknownSubsystem(f'Unknown subsystem type: {grain_type!r}')
 
     def __iter__(self):
         return iter(self._subsystems.values())