+++ cob/subsystems/models_subsystem.py
@@ -21,7 +21,7 @@ class ModelsSubsystem(SubsystemBase):
         if env_override:
             flask_app.config['SQLALCHEMY_DATABASE_URI'] = env_override
         else:
-            flask_app.config.setdefault('SQLALCHEMY_DATABASE_URI', 'sqlite:///{}'.format(os.path.join(self.project.root, '.cob', 'db.sqlite')))
+            flask_app.config.setdefault('SQLALCHEMY_DATABASE_URI', f'sqlite:///{os.path.join(self.project.root, ".cob", "db.sqlite")}')
 
         context.db = SQLAlchemy(flask_app)
         Migrate(flask_app, context.db).init_app(flask_app)