+++ cob/utils/config.py
@@ -1,21 +1,52 @@
 import copy
 import os
+from pathlib import Path
 
 import yaml
 
 
+def get_etc_config_path(project_name):
+    return Path('/etc/cob/conf.d') / project_name
+
+
+def get_config_override_locations(project_name):
+    return [
+        get_etc_config_path(project_name),
+        Path('~/.config/cob/projects').expanduser() / project_name,
+    ]
+
+
 def load_overrides(root, *, environ=None):
+
     if environ is None:
         environ = os.environ
-    config_dir = environ.get('COB_CONFIG_DIR')
-    if config_dir is not None and os.path.isdir(config_dir):
-        for yaml_file in sorted(os.listdir(config_dir)):
-            with open(os.path.join(config_dir, yaml_file)) as f:
-                root = merge_config(root, yaml.load(f))
+
+    config_dirs = []
+    project_name = root.get('name', None)
+    if project_name is not None:
+        config_dirs.extend(get_config_override_locations(project_name))
+    config_dir_env = environ.get('COB_CONFIG_DIR')
+    if config_dir_env:
+        config_dirs.append(Path(config_dir_env))
+
+    for config_dir in config_dirs:
+        if not isinstance(config_dir, Path):
+            config_dir = Path(config_dir)
+
+        if config_dir.is_dir():
+            for yaml_file in sorted(config_dir.iterdir(), key=lambda p: p.stem):
+                if yaml_file.stem.startswith(('_', '.')) or yaml_file.suffix != '.yml':
+                    continue
+                with yaml_file.open() as f:
+                    root = merge_config(root, yaml.load(f))
     return root
 
-def merge_config(parent, *children):
-    returned = copy.deepcopy(parent)
+
+def merge_config(parent, *children, in_place=False):
+    if in_place:
+        returned = parent
+    else:
+        returned = copy.deepcopy(parent)
     for child in children:
         for key, value in child.items():
             if isinstance(value, dict):
@@ -23,8 +54,8 @@ def merge_config(parent, *children):
                     returned[key] = value.copy()
                 else:
                     if not isinstance(returned[key], dict):
-                        raise ValueError('Cannot merge dictionaries: value of {!r} is not a dictionary'.format(key))
-                    returned[key].update(value)
+                        raise ValueError(f'Cannot merge dictionaries: value of {key!r} is not a dictionary')
+                    merge_config(returned[key], value, in_place=True)
             else:
                 returned[key] = copy.deepcopy(value)
     return returned