+++ tests/conftest.py
@@ -1,13 +1,30 @@
-import os
-from uuid import uuid4
-
 import logbook
-
+from pathlib import Path
 import pytest
+from uuid import uuid4
+import yaml
 
 from .project import Project
-from cob.bootstrapping import _PREVENT_REENTRY_ENV_VAR
-from cob.utils.develop import _COB_DEVELOP_MODE
+
+
+config = None
+
+
+def pytest_addoption(parser):
+    parser.addoption("--prebuilt", action="store_true", default=False,
+        help="Use prebuilt docker images")
+
+
+@pytest.fixture(autouse=True)
+def configure(request):
+    global config # pylint: disable=global-statement
+    config = request.config
+
+    @request.addfinalizer
+    def cleanup():    # pylint: disable=unused-variable
+        global config # pylint: disable=trailing-whitespace, global-statement
+        config = None
+
 
 
 @pytest.fixture(scope='session', autouse=True)
@@ -17,7 +34,7 @@ def setup_logging():
 
 @pytest.fixture
 def project_name():
-    return 'proj{}'.format(str(uuid4()).replace('-', '_'))
+    return f'proj{str(uuid4()).replace("-", "_")}'
 
 
 @pytest.fixture
@@ -25,3 +42,22 @@ def empty_project(tmpdir, project_name):
     returned = Project(tmpdir, project_name)
     returned.generate_project()
     return returned
+
+
+class _ProjectFactory:
+
+    def __init__(self, root):
+        self.root = root
+        self.config = {}
+
+    def build(self):
+        from cob.project import Project
+        with (self.root / '.cob-project.yml').open('w') as f:
+            yaml.dump(self.config, f)
+        return Project(self.root)
+
+
+@pytest.fixture
+def project_factory(tmpdir):
+    tmpdir = Path(tmpdir)
+    return _ProjectFactory(tmpdir)