+++ tests/test_basics.py
@@ -5,8 +5,9 @@ from .project import Project
 @pytest.mark.parametrize('blueprint_type', ['directory', 'file'])
 def test_blueprints(blueprint_type):
     with Project('basics').server_context() as app:
-        assert app.get('blueprints/{}/test'.format(blueprint_type)).content.decode('utf-8') == 'this is {}'.format(blueprint_type)
+        assert app.get(f'blueprints/{blueprint_type}/test').content.decode('utf-8') == f'this is {blueprint_type}'
+
 
 def test_default_config_no_track_modifications():
     with Project('basics').server_context() as app:
-        assert app.get('config').json()['SQLALCHEMY_TRACK_MODIFICATIONS'] == False
+        assert app.get('config').json()['flask_config']['SQLALCHEMY_TRACK_MODIFICATIONS'] == False