+++ tests/test_config.py
@@ -8,13 +8,14 @@ def test_merge_config():
 
 
 def test_merge_config_recursive():
-    a = {'a': 'b', 'c': {'x': 'y'}}
-    b = {'a2': 'b2', 'c': {'x2': 'y2'}}
+    a = {'a': 'b', 'c': {'x': 'y', 'd': {'z': 1}}}
+    b = {'a2': 'b2', 'c': {'x2': 'y2', 'd': {'z2': 4}}}
     result = merge_config(a, b)
     assert result == {
         'a': 'b',
         'a2': 'b2',
-        'c': {'x': 'y', 'x2': 'y2'}}
+        'c': {'x': 'y', 'x2': 'y2', 'd': {'z': 1, 'z2': 4}},
+    }
     assert result is not a
     assert result['c'] is not a['c']
 
@@ -32,8 +33,8 @@ def test_config_overrides(tmpdir, request):
     root_config = {'a': 2}
 
     for i in range(3):
-        with tmpdir.join('{:03}.yml'.format(i)).open('w') as f:
-            f.write('a: {}'.format(1000+i))
+        with tmpdir.join(f'{i:03}.yml').open('w') as f:
+            f.write(f'a: {i+1000}')
 
     cfg = load_overrides(root_config, environ={'COB_CONFIG_DIR': str(tmpdir)})
     assert cfg is not root_config